Add support for BYU SIR images.
Had to add an explicit LoadSIR function to xv.h in order for it to compile with GCC 14. I also added a config.h option. Support seems to be picky; this is worth exploring further. xvloadsir.c is also worth another look for things that were okay in 1999, but not so much now.